requirements.txt
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Faker==22.5.1
jupyter==1.0.0
ipykernel==6.29.0
